chore: convert ct-detect-third-party spec to vitest
